Accounts view, customer transactions, and employee transactions.

+++ src/main/java/com/example/generation/services/UserService.java
@@ -19,7 +19,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> findById(int id){
+    public Optional<User> findById(Long id){
         return userRepository.findById(id);
     }
 
@@ -31,7 +31,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteById(int id){
+    public void deleteById(Long id){
         userRepository.deleteById(id);
     }
 }
